turn remaining analyzers into classes
