+++ analyzers/DummyAnalyzer.cc
@@ -1,7 +1,8 @@
 #include "DummyAnalyzer.h"
 
-void DummyAnalyzer::Analyze()
+void DummyAnalyzer::Analyze(bool isData, int option, string outputFileName, string label)
 {
+    //RazorAnalyzer::Analyze(isData, option, outputFileName, label);
     if (fChain == 0) return;
 
     Long64_t nentries = fChain->GetEntriesFast();
