+++ analyzers/RazorDM.cc
@@ -1,5 +1,4 @@
-#define RazorAnalyzer_cxx
-#include "RazorAnalyzer.h"
+#include "RazorDM.h"
 #include "JetCorrectorParameters.h"
 
 //C++ includes
@@ -9,25 +8,15 @@
 
 using namespace std;
 
-#define _phodebug 0
-#define use25nsSelection 1
-
-
-struct PhotonCand
-{
-  TLorentzVector pho;
-  bool passIso;
-  bool passEleVeto;
-};
-
-void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool isData )
+void RazorDM::Analyze(bool isData, int option, string outFileName, string label)
 {
   //initialization: create one TTree for each analysis box 
   cout << "Initializing..." << endl;
+  bool combineTrees = true;
   if ( outFileName.empty() )
     {
-      cout << "RazorPhotonDM: Output filename not specified!" << endl << "Using default output name RazorPhotonDM.root" << endl;
-      outFileName = "RazorPhotonDM.root";
+      cout << "RazorDM: Output filename not specified!" << endl << "Using default output name RazorDM.root" << endl;
+      outFileName = "RazorDM.root";
     }
   if (combineTrees) cout << "Using combineTrees" << endl;
   else cout << "Using razorBoxes" << endl;
@@ -68,7 +57,7 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
   FactorizedJetCorrector *JetCorrector = new FactorizedJetCorrector( correctionParameters );
   
   //one tree to hold all events
-  TTree *razorTree = new TTree("RazorPhotonDM", "Info on selected razor DM events");
+  TTree *razorTree = new TTree("RazorDM", "Info on selected razor DM events");
   
   // tree to compute the cut efficiency
   //TTree *effTree = new TTree("CutEfficiency","Efficiencies of cuts on DM events"); 
@@ -91,21 +80,19 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
   TH1F *NEvents = new TH1F("NEvents", "NEvents", 1, 1, 2);
   
   //tree variables
-  int nPho;
-  float PhoPt[50], PhoEta[50], PhoPhi[50];
-  bool PhoPassIso[50], PhoPassEleVeto[50];
   int nSelectedJets, nBTaggedJetsL, nBTaggedJetsM, nBTaggedJetsT;
   int nLooseMuons, nTightMuons, nLooseElectrons, nTightElectrons, nTightTaus;
   UInt_t run, lumi, event;
   float MuonPt[5], MuonEta[5], MuonPhi[5], MuonE[5];
   float ElePt[5], EleEta[5], ElePhi[5], EleE[5];
-  float MR;
+  float MR, deltaPhi;
   float HT, MHT;
   float alphaT, dPhiMin;
   float Rsq, t1Rsq, RsqCorr, t1RsqCorr;
   float t1metPt, t1metPhi, metPtCorr, metPhiCorr, t1metPtCorr, t1metPhiCorr;
   RazorBox box;
   float JetPt_uncorr[30], JetEta_uncorr[30], JetPhi_uncorr[30], JetE_uncorr[30];
+  float LeadJetNeutralHadronFraction, LeadJetChargedHadronFraction;
   float JetPt[30], JetEta[30], JetPhi[30], JetE[30];
   bool hasMatchingGenJet[30];
   int matchingGenJetIndex[30];
@@ -115,17 +102,10 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
     //set branches on big tree
   if(combineTrees)
     {
-      razorTree->Branch("run",&run,"run/i");
-      razorTree->Branch("lumi",&lumi,"lumi/i");
-      razorTree->Branch("event",&event,"event/i");
-      
-      razorTree->Branch("nPho", &nPho, "nPho/i");
-      razorTree->Branch("PhoPt", PhoPt, "PhoPt[nPho]/F");
-      razorTree->Branch("PhoEta", PhoEta, "PhoEta[nPho]/F");
-      razorTree->Branch("PhoPhi", PhoPhi, "PhoPhi[nPho]/F");
-      razorTree->Branch("PhoPassIso", PhoPassIso, "PhoPassIso[nPho]/O");
-      razorTree->Branch("PhoPassEleVeto", PhoPassEleVeto, "PhoPassEleVeto[nPho]/O");
-      
+    razorTree->Branch("run",&run,"run/i");
+    razorTree->Branch("lumi",&lumi,"lumi/i");
+    razorTree->Branch("event",&event,"event/i");
+   
       razorTree->Branch("nSelectedJets", &nSelectedJets, "nSelectedJets/I");
       razorTree->Branch("nBTaggedJetsL", &nBTaggedJetsL, "nBTaggedJetsL/I");
       razorTree->Branch("nBTaggedJetsM", &nBTaggedJetsM, "nBTaggedJetsM/I");
@@ -145,6 +125,8 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
       razorTree->Branch("JetPt", JetPt, "JetPt[nSelectedJets]/F");
       razorTree->Branch("JetEta", JetEta, "JetEta[nSelectedJets]/F");
       razorTree->Branch("JetPhi", JetPhi, "JetPhi[nSelectedJets]/F");
+      razorTree->Branch("LeadJetNeutralHadronFraction", &LeadJetNeutralHadronFraction, "LeadJetNeutralHadronFraction/F"); 
+      razorTree->Branch("LeadJetChargedHadronFraction", &LeadJetChargedHadronFraction, "LeadJetChargedHadronFraction/F"); 
       
       razorTree->Branch("alphaT", &alphaT, "alphaT/F");
       razorTree->Branch("dPhiMin", &dPhiMin, "dPhiMin/F");
@@ -152,6 +134,7 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
       razorTree->Branch("MR", &MR, "MR/F");
       razorTree->Branch("Rsq", &Rsq, "Rsq/F");
       razorTree->Branch("t1Rsq", &t1Rsq, "t1Rsq/F");
+      razorTree->Branch("deltaPhi", &deltaPhi, "deltaPhi/F");
       razorTree->Branch("metPt", &metPt, "metPt/F");
       razorTree->Branch("metPhi", &metPhi, "metPhi/F");
       razorTree->Branch("t1metPt", &t1metPt, "t1metPt/F");
@@ -266,11 +249,10 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
         event = eventNum;
 
         //reset tree variables
-	nPho = 0;
         nSelectedJets = 0;
         nBTaggedJetsL = 0;
-	nBTaggedJetsM = 0;
-	nBTaggedJetsT =0;
+	    nBTaggedJetsM = 0;
+	    nBTaggedJetsT =0;
         nLooseMuons = 0;
         nTightMuons = 0;
         nLooseElectrons = 0;
@@ -280,13 +262,16 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
         dPhiMin   = -1.0;
         MR        = -1.0;
         Rsq       = -1.0;
-	t1Rsq     = -1.0;
-	RsqCorr   = -1.0;
-	t1RsqCorr = -1.0;
-	leadingJetPt = -1;
-	leadingJetEta = 0.0;
-	subLeadingJetPt = -1;
-	
+	    deltaPhi     = -1.0;
+	    t1Rsq     = -1.0;
+	    RsqCorr   = -1.0;
+	    t1RsqCorr = -1.0;
+      leadingJetPt = -1;
+      leadingJetEta = 0.0;
+      subLeadingJetPt = -1;
+	    LeadJetChargedHadronFraction = -1.; 
+	    LeadJetNeutralHadronFraction = -1.; 
+	  
 	for ( int j = 0; j < 30; j++ )
           {
 	    JetE_uncorr[j]   = 0.0;
@@ -298,19 +283,20 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	    JetPhi[j] = 0.0;
 	    JetEta[j] = 0.0;
           }
-	
-	for ( int j = 0; j < 50; j++ )
-	  {
-	    PhoPt[j]  = -999;
-	    PhoEta[j] = -999;
-	    PhoPhi[j] = -999;
-	    PhoPassIso[j] = false;
-	    PhoPassEleVeto[j] = false;
-	  }
-	
         if(combineTrees) box = NONE;
 
-	//TODO: triggers!
+	//DEBUGGING GENJETS OUTPUT
+	/*	if (jentry < 5){
+	  cout << "nGenJets: " << nGenJets << endl;
+	  for (int j = 0; j < nGenJets; j++){
+	    cout << "genJetE: " << genJetE[j] << endl;
+	    cout << "genJetPt: " << genJetPt[j] << endl;
+	    cout << "genJetEta: " << genJetEta[j] << endl;
+	    cout << "genJetPhi: " << genJetPhi[j] << endl;
+	  }
+	}
+	*/
+        //TODO: triggers!
         bool passedLeptonicTrigger = true;
         bool passedHadronicTrigger= true;
         if ( !(passedLeptonicTrigger || passedHadronicTrigger) ) continue; //ensure event passed a trigger
@@ -328,7 +314,7 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	  if( isTightMuon(i) ) nTightMuons++;
         }
 	
-	
+
 	//------------------------------
 	// Reco+ID Muons
 	//------------------------------
@@ -356,89 +342,12 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	
         
 	//------------------------
-	//Photon ID + Selection
-	//------------------------
-	//photon selection
-	std::vector<TLorentzVector> GoodPhotons;
-	std::vector< PhotonCand > GoodPhoCand;
-	int nPhotonsAbove40GeV = 0;
-	for(int i = 0; i < nPhotons; i++)
-	  {
-	    //ID cuts -- apply isolation after candidate pair selection
-	    if ( _phodebug ) std::cout << "pho# " << i << " phopt1: " << phoPt[i] << " pho_eta: " << phoEta[i] << std::endl;
-	    if ( !photonPassLooseIDWithoutEleVeto(i,use25nsSelection) ) 
-	      {
-		if ( _phodebug ) std::cout << "[DEBUG]: failed run2 ID" << std::endl;
-		continue;
-	      }
-
-	    //only isolated photons
-	    if ( !photonPassLooseIso(i,use25nsSelection) ) continue;
-	    
-	    //Defining Corrected Photon momentum
-	    //float pho_pt = phoPt[i];//nominal pt
-	    float pho_pt_corr = pho_RegressionE[i]/cosh(phoEta[i]);//regression corrected pt
-	    TVector3 vec;
-	    //vec.SetPtEtaPhi( pho_pt, phoEta[i], phoPhi[i] );
-	    vec.SetPtEtaPhi( pho_pt_corr, phoEta[i], phoPhi[i] );
-	    
-	    if ( phoPt[i] < 20.0 )
-	      {
-		if ( _phodebug ) std::cout << "[DEBUG]: failed pt" << std::endl;
-		continue;
-	      }
-	    
-	    if( fabs(pho_superClusterEta[i]) > 2.5 )
-	      {
-		//allow photons in the endcap here, but if one of the two leading photons is in the endcap, reject the event
-		if ( _phodebug ) std::cout << "[DEBUG]: failed eta" << std::endl;
-		continue; 
-	      }
-	    
-	    if ( fabs(pho_superClusterEta[i]) > 1.4442 && fabs(pho_superClusterEta[i]) < 1.566 )
-	      {
-		//Removing gap photons
-		if ( _phodebug ) std::cout << "[INFO]: failed gap" << std::endl;
-		continue;
-	      }
-	    //photon passes
-	    if( phoPt[i] > 34.0 ) nPhotonsAbove40GeV++;
-	    //setting up photon 4-momentum with zero mass
-	    TLorentzVector thisPhoton;
-	    thisPhoton.SetVectM( vec, .0 );
-	    GoodPhotons.push_back( thisPhoton );
-	    PhotonCand mypho;
-	    mypho.pho = thisPhoton;
-	    mypho.passIso = photonPassLooseIso(i,use25nsSelection);
-	    mypho.passEleVeto = pho_passEleVeto[i];
-	    GoodPhoCand.push_back( mypho );
-	    nPho++;
-	  }
-	
-	if ( nPho < 1 ) continue;
-	//sort photons
-	auto sortVLV = []( TLorentzVector a, TLorentzVector b ){ return a.Pt() > b.Pt() ?  true : false; };
-	auto sortPhotonCand = [] ( PhotonCand a, PhotonCand b ){ return a.pho.Pt() > b.pho.Pt() ? true : false; };
-	std::sort( GoodPhotons.begin(), GoodPhotons.end(), sortVLV );
-	std::sort( GoodPhoCand.begin(), GoodPhoCand.end(), sortPhotonCand );
-	
-	int iPho = 0;
-	for( auto& tmp : GoodPhoCand )
-	  {
-	    PhoPt[iPho] = tmp.pho.Pt();
-	    PhoEta[iPho] = tmp.pho.Eta();
-	    PhoPhi[iPho] = tmp.pho.Phi();
-	    PhoPassIso[iPho] = tmp.passIso;
-	    PhoPassEleVeto[iPho] = tmp.passEleVeto;
-	    iPho++;
-	  }
-
-	
-	//------------------------
 	//Jet ID + Correction
 	//------------------------
 	vector<TLorentzVector> GoodJets;
 	vector<TLorentzVector> GoodJets_uncorr;
+    vector<float> JetChargedHadronFraction;
+    vector<float> JetNeutralHadronFraction;
         int numJetsAbove80GeV = 0;
 	for( int i = 0; i < nJets; i++ )
 	  {
@@ -480,39 +389,32 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	      }
 	    if ( deltaR > 0.0 && deltaR < 0.3 ) continue; //jet matches a Loose Electron
 	    
-	    deltaR = -1.0;
-	    for ( auto& pho : GoodPhotons )
-	      {
-		double thisDR = thisJet_uncorr.DeltaR(pho);
-		if(deltaR < 0.0 || thisDR < deltaR) deltaR = thisDR;
-	      }
-	     if ( deltaR > 0.0 && deltaR < 0.5 ) continue; //jet matches a ID photon
-	     
-	     if ( thisJet.Pt() > 80.0 ) numJetsAbove80GeV++;
+	    if ( thisJet.Pt() > 80.0 ) numJetsAbove80GeV++;
 	    GoodJets_uncorr.push_back(thisJet_uncorr);
 	    
 	    GoodJets.push_back(thisJet);
-	    nSelectedJets++;
+        JetNeutralHadronFraction.push_back(jetNeutralHadronEnergyFraction[i]);
+        JetChargedHadronFraction.push_back(jetChargedHadronEnergyFraction[i]);
+	    
+        nSelectedJets++;
 	    //b-tagging 
 	    if(isCSVL(i)) nBTaggedJetsL++;
 	    if(isCSVM(i)) nBTaggedJetsM++;
 	    if(isCSVT(i)) nBTaggedJetsT++;
 	  }
 	
-        if ( numJetsAbove80GeV < 1 ) continue; //event fails to have two 80 GeV jets
-	//sort Jets
-	std::sort( GoodJets.begin(), GoodJets.end(), sortVLV );
+        if ( numJetsAbove80GeV < 2 ) continue; //event fails to have two 80 GeV jets
 	
-	int jIndex = 0;
-	for (auto& tmpJet: GoodJets_uncorr)
-	  {
+	    int jIndex = 0;
+	    for (auto& tmpJet: GoodJets_uncorr)
+	    {
 	    JetPt_uncorr[jIndex] = tmpJet.Pt();
 	    JetE_uncorr[jIndex] = tmpJet.E();
 	    JetPhi_uncorr[jIndex] = tmpJet.Phi();
 	    JetEta_uncorr[jIndex] = tmpJet.Eta();
 	    jIndex=jIndex+1;
-	  }  
-	
+	    }  
+
 	//Find whether there is a matching genJet for the reconstructed jet
 	jIndex = 0;
 	for ( auto& tmpJet : GoodJets )
@@ -528,7 +430,7 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	      }
 	    jIndex++;
 	  }
-	
+
 	//output information for corrected jets
 	jIndex = 0;
 	for (auto& tmpJet: GoodJets)
@@ -537,41 +439,52 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	    JetE[jIndex]   = tmpJet.E();
 	    JetEta[jIndex] = tmpJet.Eta();
 	    JetPhi[jIndex] = tmpJet.Phi();
+	
+        if (JetPt[jIndex]> leadingJetPt && JetPt[jIndex]<14000) 
+        {
+	    subLeadingJetPt = leadingJetPt;
+	    leadingJetPt = JetPt[jIndex];
+        leadingJetEta = JetEta[jIndex];
+        LeadJetNeutralHadronFraction = JetNeutralHadronFraction.at(jIndex);
+        LeadJetChargedHadronFraction = JetChargedHadronFraction.at(jIndex);
+	    }
+	    else if (JetPt[jIndex]>subLeadingJetPt && JetPt[jIndex]<14000) 
+        {
+	     subLeadingJetPt = JetPt[jIndex];
+	    }
+        auto sortJets = []( TLorentzVector a, TLorentzVector b ){ return a.Pt() > b.Pt() ? true : false; };
+        std::sort( GoodJets.begin() , GoodJets.end(), sortJets);
+	    
 	    jIndex++;
 	  }
+    
+    // Compute the variables alpha T and dPhiMin using the selected jets
+    alphaT = GetAlphaT(GoodJets);
+    dPhiMin = GetDPhiMin(GoodJets);  
 
-	//Create Photon+Jet Collection
-	std::vector<TLorentzVector> PhotonsAndJets;
-	for( auto& tmp : GoodPhotons ) PhotonsAndJets.push_back( tmp );
-	for( auto& tmp : GoodJets ) PhotonsAndJets.push_back( tmp );
-	
-	// Compute the variables alpha T and dPhiMin using the selected jets
-	alphaT = GetAlphaT(PhotonsAndJets);
-	dPhiMin = GetDPhiMin(PhotonsAndJets);  
-	
 	//Compute the razor variables using the selected jets and possibly leptons
 	TLorentzVector PFMET = makeTLorentzVectorPtEtaPhiM(metPt, 0, metPhi, 0);
 	TLorentzVector t1PFMET = makeTLorentzVectorPtEtaPhiM( metType1Pt, 0, metType1Phi, 0 );
 	t1metPt  = metType1Pt;
 	t1metPhi = metType1Phi;
 	
-	vector<TLorentzVector> hemispheres = getHemispheres( PhotonsAndJets );
-	
-	MR    = computeMR(hemispheres[0], hemispheres[1]); 
+    vector<TLorentzVector> hemispheres = getHemispheres( GoodJets );
+    
+    MR    = computeMR(hemispheres[0], hemispheres[1]); 
 	Rsq   = computeRsq(hemispheres[0], hemispheres[1], PFMET);
 	t1Rsq = computeRsq(hemispheres[0], hemispheres[1], t1PFMET);
-	
+    deltaPhi = fabs(hemispheres[0].DeltaPhi(hemispheres[1])); 
 
-	// Compute HT and MHT
-	float MhtX = 0., MhtY = 0.;
-	HT = 0.; 
-	for (auto& obj : PhotonsAndJets) { HT += obj.Pt(); MhtX += obj.Px(); MhtY += obj.Py(); }
-	
-	TLorentzVector MyMHT;
-	MyMHT.SetPxPyPzE(-MhtX, -MhtY, 0, sqrt(pow(MhtX,2) + pow(MhtY,2)));
-	
-	MHT = MyMHT.Pt();
-	
+    // Compute HT and MHT
+    float MhtX = 0., MhtY = 0.;
+    HT = 0.; 
+    for (auto& obj : GoodJets) { HT += obj.Pt(); MhtX += obj.Px(); MhtY += obj.Py(); }
+
+      TLorentzVector MyMHT;
+      MyMHT.SetPxPyPzE(-MhtX, -MhtY, 0, sqrt(pow(MhtX,2) + pow(MhtY,2)));
+
+      MHT = MyMHT.Pt();
+    
 	//MuMu Box
         if ( passedLeptonicTrigger && nLooseMuons > 1 && nLooseElectrons == 0 && nBTaggedJetsL == 0 )
 	  {
@@ -725,7 +638,8 @@ void RazorAnalyzer::RazorPhotonDM(string outFileName, bool combineTrees, bool is
 	      razorTree->Fill();
 	    }
 	    else razorBoxes["OneBJet"]->Fill();
-	  }       
+	  } 
+      
     }//end of event loop
     
     cout << "Writing output trees..." << endl;
