+++ include/AnalyzerTemplate.txt
@@ -6,7 +6,7 @@
 class %ANALYZER%: public RazorAnalyzer {
     public: 
         %ANALYZER%(TTree *tree=0): RazorAnalyzer(tree) { }
-        void Analyze();
+        void Analyze(bool isData, int option, string outputFileName, string label);
 };
 
 #endif
