+++ src/RunAnalyzerTemplate.txt
@@ -170,7 +170,7 @@ int main(int argc, char* argv[]){
     //------ EXECUTE ------//
     cout << "Executing %ANALYZER%..." << endl;
     analyzer.EnableAll();
-    analyzer.Analyze();
+    analyzer.Analyze(isData, option, outputFileName, label);
     cout << "Process completed!" << endl;
     cerr << "------------------------------" << endl; //present so that an empty .err file corresponds to a failed job
